Rename utils to utils/helper
Refactoring all page calls
