+++ app/utils/helpers.ts
@@ -1,0 +1,4 @@
+//
+export function isDev(): boolean {
+  return process.env.NODE_ENV === 'development'
+}
